requirements.txt
streamlit
opencv-python-headless
numpy
ultralytics
torch
torchvision
timm
plotly
av
Pillow
